extract scraper into its own class
